move extend sos coefficient func to filter classes
